update generated files, add and configure media setup/admin

+++ components/model/library/ModelLibraryAdmin.vue
@@ -184,9 +184,11 @@
       msgDeleteSuccess: string;
       msgDeleteError: string;
       actionConfigs: Record<string, ActionConfig>;
+      canAdd: (obj: MobilettoOrmObject, objList: MobilettoOrmObject[]) => boolean;
       canEdit: (obj: MobilettoOrmObject, objList: MobilettoOrmObject[]) => boolean;
       canDelete: (obj: MobilettoOrmObject, objList: MobilettoOrmObject[]) => boolean;
       deleteConfirmationMessage: string;
+      maxDeleteConfirmations: number;
     }>(),{
       labelPrefixes: () => ["label_", ""],
       typeNameMessage: () => "typename_library",
@@ -197,8 +199,10 @@
       msgDeleteSuccess: () => "admin_info_deleted",
       msgDeleteError: () => "admin_info_delete_error",
       actionConfigs: () => ({}),
+      canAdd: () => true,
       canEdit: () => true,
       canDelete: () => true,
+      maxDeleteConfirmations: () => 3,
     },
   );
 
@@ -463,7 +467,18 @@
     }
     return true;
   };
+  const delConfirmCount = ref(0);
+  const deletingObject = ref(null);
   const delObject = (obj: MobilettoOrmObject) => {
+      if (props.deleteConfirmationMessage && props.deleteConfirmationMessage.length > 0 && delConfirmCount.value < maxDeleteConfirmations) {
+          if (confirm(props.deleteConfirmationMessage)) {
+              delConfirmCount.value = delConfirmCount.value + 1;
+          } else {
+              delConfirmCount.value = 0;
+              return;
+          }
+      }
+      deletingObject.value = obj;
       const id = LibraryTypeDef.id(obj);
       libraryStore.delete(id, deleteLibraryServerErrors)
           .then(() => {
@@ -479,7 +494,7 @@
     if (deleteLibraryServerErrors.value && Object.keys(deleteLibraryServerErrors.value).length > 0) {
         errorSnackbar.value = parseMessage(props.msgDeleteError, messages.value, {
             type: messages.value.typename_library,
-            id: LibraryTypeDef.id(obj),
+            id: deletingObject.value ? LibraryTypeDef.id(deletingObject.value) : "null",
             error: JSON.stringify(deleteLibraryServerErrors),
         });
     }
