+++ components/model/mediaProfile/ModelMediaProfileAdmin.vue
@@ -183,9 +183,11 @@
       msgDeleteSuccess: string;
       msgDeleteError: string;
       actionConfigs: Record<string, ActionConfig>;
+      canAdd: (obj: MobilettoOrmObject, objList: MobilettoOrmObject[]) => boolean;
       canEdit: (obj: MobilettoOrmObject, objList: MobilettoOrmObject[]) => boolean;
       canDelete: (obj: MobilettoOrmObject, objList: MobilettoOrmObject[]) => boolean;
       deleteConfirmationMessage: string;
+      maxDeleteConfirmations: number;
     }>(),{
       labelPrefixes: () => ["label_", ""],
       typeNameMessage: () => "typename_mediaProfile",
@@ -196,8 +198,10 @@
       msgDeleteSuccess: () => "admin_info_deleted",
       msgDeleteError: () => "admin_info_delete_error",
       actionConfigs: () => ({}),
+      canAdd: () => true,
       canEdit: () => true,
       canDelete: () => true,
+      maxDeleteConfirmations: () => 3,
     },
   );
 
@@ -437,7 +441,18 @@
     }
     return true;
   };
+  const delConfirmCount = ref(0);
+  const deletingObject = ref(null);
   const delObject = (obj: MobilettoOrmObject) => {
+      if (props.deleteConfirmationMessage && props.deleteConfirmationMessage.length > 0 && delConfirmCount.value < maxDeleteConfirmations) {
+          if (confirm(props.deleteConfirmationMessage)) {
+              delConfirmCount.value = delConfirmCount.value + 1;
+          } else {
+              delConfirmCount.value = 0;
+              return;
+          }
+      }
+      deletingObject.value = obj;
       const id = MediaProfileTypeDef.id(obj);
       mediaProfileStore.delete(id, deleteMediaProfileServerErrors)
           .then(() => {
@@ -453,7 +468,7 @@
     if (deleteMediaProfileServerErrors.value && Object.keys(deleteMediaProfileServerErrors.value).length > 0) {
         errorSnackbar.value = parseMessage(props.msgDeleteError, messages.value, {
             type: messages.value.typename_mediaProfile,
-            id: MediaProfileTypeDef.id(obj),
+            id: deletingObject.value ? MediaProfileTypeDef.id(deletingObject.value) : "null",
             error: JSON.stringify(deleteMediaProfileServerErrors),
         });
     }
