+++ components/model/profileJob/ModelProfileJobAdmin.vue
@@ -182,9 +182,11 @@
       msgDeleteSuccess: string;
       msgDeleteError: string;
       actionConfigs: Record<string, ActionConfig>;
+      canAdd: (obj: MobilettoOrmObject, objList: MobilettoOrmObject[]) => boolean;
       canEdit: (obj: MobilettoOrmObject, objList: MobilettoOrmObject[]) => boolean;
       canDelete: (obj: MobilettoOrmObject, objList: MobilettoOrmObject[]) => boolean;
       deleteConfirmationMessage: string;
+      maxDeleteConfirmations: number;
     }>(),{
       labelPrefixes: () => ["label_", ""],
       typeNameMessage: () => "typename_profileJob",
@@ -195,8 +197,10 @@
       msgDeleteSuccess: () => "admin_info_deleted",
       msgDeleteError: () => "admin_info_delete_error",
       actionConfigs: () => ({}),
+      canAdd: () => true,
       canEdit: () => true,
       canDelete: () => true,
+      maxDeleteConfirmations: () => 3,
     },
   );
 
@@ -405,7 +409,18 @@
     }
     return true;
   };
+  const delConfirmCount = ref(0);
+  const deletingObject = ref(null);
   const delObject = (obj: MobilettoOrmObject) => {
+      if (props.deleteConfirmationMessage && props.deleteConfirmationMessage.length > 0 && delConfirmCount.value < maxDeleteConfirmations) {
+          if (confirm(props.deleteConfirmationMessage)) {
+              delConfirmCount.value = delConfirmCount.value + 1;
+          } else {
+              delConfirmCount.value = 0;
+              return;
+          }
+      }
+      deletingObject.value = obj;
       const id = ProfileJobTypeDef.id(obj);
       profileJobStore.delete(id, deleteProfileJobServerErrors)
           .then(() => {
@@ -421,7 +436,7 @@
     if (deleteProfileJobServerErrors.value && Object.keys(deleteProfileJobServerErrors.value).length > 0) {
         errorSnackbar.value = parseMessage(props.msgDeleteError, messages.value, {
             type: messages.value.typename_profileJob,
-            id: ProfileJobTypeDef.id(obj),
+            id: deletingObject.value ? ProfileJobTypeDef.id(deletingObject.value) : "null",
             error: JSON.stringify(deleteProfileJobServerErrors),
         });
     }
